feat: simplify code and add padding
