+++ src/i18n/index.ts
@@ -1,5 +1,4 @@
 import { loadTranslations } from '@okp4/ui'
-
 import footer_en from './footer/footer_en.json'
 import footer_fr from './footer/footer_fr.json'
 import dashboard_en from './dashboard/dashboard_en.json'
