+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app'
 import dynamic from 'next/dynamic'
-import { Layout } from '../components/layout/Layout'
+import { Layout } from '../components'
 import './styles.scss'
 
 const ThemeProvider = dynamic(async () => (await import('@okp4/ui')).ThemeProvider, {
