+++ src/pages/home/home.scss
@@ -5,22 +5,16 @@
     height: 100%;
     display: grid;
     grid-template-columns: repeat(4, 1fr);
-    padding: 70px 130px 25px 130px;
     gap: 100px;
 
     @media screen and (max-width: 1300px) {
       grid-template-columns: repeat(2, 1fr);
-      padding: 50px 90px 25px 90px;
       gap: 50px;
     }
 
     @media screen and (max-width: 995px) {
       grid-template: auto 1fr /1fr;
-      padding: 50px 70px 25px 70px;
       gap: 50px 0;
-    }
-    @media screen and (max-width: 480px) {
-      padding: 25px 20px 0 20px;
     }
 
     .okp4-body-dashboard {
